fix totem effect remove bug
